requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langfuse>=2.57.0
openai>=1.50.0
prometheus-client==0.21.0
langtrace-python-sdk==3.8.21
